backend/requirements.txt
# backend/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# For CONDA + PyTorch already installed setup:
# 1. conda activate workspacevision
# 2. pip install git+https://github.com/facebookresearch/sam2.git
# 3. pip install -r backend/requirements.txt
#
# PyTorch/torchvision/torchaudio EXCLUDED — managed by conda.
# Installing them via pip would overwrite GPU version with CPU version.
# ─────────────────────────────────────────────────────────────────────────────
# ── CRITICAL VERSION PINS — DO NOT CHANGE WITHOUT TESTING ────────────────────
# numpy 2.x     → breaks langchain, mlflow, streamlit, pyarrow
# packaging 25+ → breaks langchain-core, mlflow, streamlit
# transformers 5.x → breaks sentence-transformers + tokenizers==0.19.1
numpy==1.26.4
packaging==24.2
transformers==4.41.2
tokenizers==0.19.1
huggingface-hub==0.23.4

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1

# ── HuggingFace (Model Hub) ───────────────────────────────────────────────────
transformers==4.41.2
accelerate==0.30.0
huggingface-hub==0.23.4
safetensors==0.4.3
timm==1.0.3

# ── Computer Vision & Image Processing ───────────────────────────────────────
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4

# ── LangChain (RAG Pipeline) ──────────────────────────────────────────────────
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
langchain-chroma==0.1.3

# ── Vector Database ───────────────────────────────────────────────────────────
chromadb>=0.4.0,<0.6.0

# ── PDF Processing ────────────────────────────────────────────────────────────
pypdf==4.2.0
tiktoken==0.7.0 

# ── OpenAI ───────────────────────────────────────────────────────────────────
openai>=1.40.0,<2.0.0

# ── Sentence Embeddings ───────────────────────────────────────────────────────
sentence-transformers==3.0.1

# ── MLflow (Experiment Tracking) ─────────────────────────────────────────────
mlflow==2.13.2

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.27.0
aiofiles==23.2.1

# ── Utilities ─────────────────────────────────────────────────────────────────
python-jose==3.3.0
python-dateutil==2.9.0
tqdm==4.66.4
